requirements.txt
praw==7.7.1
pandas==2.1.0
numpy==1.24.3
textblob==0.17.1
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
gensim==4.3.1
python-dotenv==1.0.0
folium==0.14.0
geopy==2.3.0
